requirements.txt
dash
dash_core_components
dash_html_components
dash-daq==0.5.0
flask
sqlalchemy
flask_login
werkzeug
flask_sqlalchemy
gunicorn
numpy
pandas
mysql-connector-python
matplotlib
dlib
Pillow
cmake
face-recognition
face-recognition-models
Pyrebase
firebase
python-dotenv
scikit-learn==0.24.2
mlxtend==0.19.0
scikit-image==0.17.2
opencv-python
openpyxl